% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Just import some function from RProtoBuf}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{a greeting}
}
\value{
logical value indicating consistent serialization/deserialization
}
\description{
Import some function from RProtoBuf
for distributed computations
}

